adapted to node, as to enable out big file
we couldn’t convert the other 3 big .in files lines into variables in
the last format, so needed to do this. still, kittens was left out and
the out files came with an extra blank line in the end. the final
solution was submitted removing those by hand, after the online
dashboard warned us about it.